ivy: Add a sort-matches-function to raise full matches when asked
In Ivy, there are two types of sorting functions: those in
ivy-sort-functions-alist and those in
ivy-sort-matches-functions-alist. The former only runs once when
entering completion, while the latter runs on every search update and
has access to the current search term.

To implement prescient-sort-full-matches-first, the latter needs to be
used. This commit adds a function appropriate for the
ivy-sort-matches-functions-alist, but does not yet enable it
automatically when turning on ivy-prescient-mode.